# src/main/docs/api.yaml
openapi: 3.1.0
info:
  title: Agricultural federation API
  description: |
    # Versioning
    This API uses semantic versioning (MAJOR.MINOR.PATCH). Available versions are : 
    - v0.0.1 (2026-04-21) : POST /collectivities for collectivities creation and POST /members for members creation.
    - v0.0.2 (2026-04-22) : POST /collectivities/{collectivityId}/identifiers for assigning unique number and name to collectivities.
  version: 0.0.2
servers:
  - url: 'https://localhost:8080'
paths:
  '/collectivities':
    post:
      summary: Create list of collectivities
      description: |
        During collectivity request body creation, only members ID are provided to identify members.
        But the response content includes members information, not only their IDs.
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreateCollectivity'
      responses:
        201:
          description: List of created collectivities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Collectivity'
        400:
          description: Collectivity without federation approval or structure missing.
        404:
          description: Member not found.

  '/collectivities/{collectivityId}/identifiers':
    post:
      summary: Assign unique number and name to a collectivity
      description: |
        Assign a unique number and name to a collectivity by the federation.
        If the proposed name already exists in the system, the request will be rejected.
      parameters:
        - name: collectivityId
          in: path
          required: true
          schema:
            type: string
        - name: X-Federation-Id
          in: header
          required: true
          schema:
            type: string
          description: Federation identifier authorizing the identifier assignment.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectivityIdentifierRequest'
      responses:
        201:
          description: Identifier assigned successfully
          headers:
            Location:
              schema:
                type: string
              description: URL of the updated collectivity with new identifiers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collectivity'
        400:
          description: |
            - Collectivity already has identifiers assigned.
            - Name already exists in the system.
            - Missing federation authorization.
        404:
          description: Collectivity not found.
  '/members':
    post:
      summary: Create list of members
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreateMember'
      responses:
        201:
          description: List of created members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Member'
        400:
          description: |
            The following case must return a bad request exception : 
            - Member with bad referees or without proper payment.
            - Membership dues not paid or registration fee not paid.
        404:
          description: Either collectivity or member not found.
components:
  schemas:
    CreateCollectivity:
      type: object
      properties:
        location:
          type: string
        members:
          type: array
          items:
            schema:
              $ref: '#/components/schemas/MemberIdentifier'
        federationApproval:
          type: boolean
        structure:
          $ref: '#/components/schemas/CreateCollectivityStructure'
    CreateCollectivityStructure:
      type: object
      properties:
        president:
          $ref: '#/components/schemas/MemberIdentifier'
        vicePresident:
          $ref: '#/components/schemas/MemberIdentifier'
        treasurer:
          $ref: '#/components/schemas/MemberIdentifier'
        secretary:
          $ref: '#/components/schemas/MemberIdentifier'
    CollectivityStructure:
      type: object
      properties:
        president:
          $ref: '#/components/schemas/Member'
        vicePresident:
          $ref: '#/components/schemas/Member'
        treasurer:
          $ref: '#/components/schemas/Member'
        secretary:
          $ref: '#/components/schemas/Member'
    Collectivity:
      type: object
      properties:
        id:
          type: string
        number:
          type: string
          description: Unique identifier number assigned by the federation
        name:
          type: string
          description: Unique name assigned by the federation
        location:
          type: string
        structure:
          $ref: '#/components/schemas/CollectivityStructure'
        members:
          type: array
          items:
            $ref: '#/components/schemas/Member'
    CollectivityIdentifierRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Proposed unique name for the collectivity
    MemberIdentifier:
      type: string
    CreateMember:
      type: object
      allOf:
        - $ref: '#/components/schemas/MemberInformation'
      properties:
        collectivityIdentifier:
          type: string
        referees:
          type: array
          items:
            $ref: '#/components/schemas/MemberIdentifier'
        registrationFeePaid:
          type: boolean
        membershipDuesPaid:
          type: boolean
    MemberInformation:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date
        gender:
          $ref: '#/components/schemas/Gender'
        address:
          type: string
        profession:
          type: string
        phoneNumber:
          type: int
        email:
          type: string
        occupation:
          $ref: '#/components/schemas/MemberOccupation'
    Member:
      type: object
      allOf:
        - $ref: '#/components/schemas/MemberInformation'
      properties:
        id:
          $ref: '#/components/schemas/MemberIdentifier'
        referees:
          type: array
          items:
            $ref: '#/components/schemas/Member'
    Gender:
      type: string
      enum:
        - MALE
        - FEMALE
    MemberOccupation:
      type: string
      enum:
        - JUNIOR
        - SENIOR
        - SECRETARY
        - TREASURER
        - VICE_PRESIDENT
        - PRESIDENT